requirements.txt
python-telegram-bot==20.7
ebooklib==0.17.1
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
Pillow==10.1.0 